+++ NostelgiAlbum/AlbumScreenViewController/AlbumScreenViewController.swift
@@ -2,14 +2,13 @@ import UIKit
 import RealmSwift
 
 class AlbumScreenViewController: UIViewController {
-    
+    // MARK: - Properties
     let realm = try! Realm()
     var pageNum : Int = 0
     var coverIndex : Int = 0
-    
-    // collectionView setting
     @IBOutlet weak var collectionView: UICollectionView!
     
+    // MARK: - View Life Cycle
     override func viewDidLoad() {
         super.viewDidLoad()
         collectionView.dataSource = self
@@ -39,7 +38,6 @@ class AlbumScreenViewController: UIViewController {
         let swipeLeft = UISwipeGestureRecognizer(target: self, action: #selector(AlbumScreenViewController.respondToSwipeGesture(_:)))
         swipeLeft.direction = UISwipeGestureRecognizer.Direction.left
         self.view.addGestureRecognizer(swipeLeft)
-        
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -47,7 +45,7 @@ class AlbumScreenViewController: UIViewController {
         navigationController?.isToolbarHidden = false
     }
     
-    // Set toolBar
+    // MARK: - Methods
     private func makeToolbarItems() -> [UIBarButtonItem]{
         let searchButton = UIBarButtonItem(image: UIImage(systemName: "magnifyingglass")?.withRenderingMode(.alwaysTemplate), style: .plain, target: self, action: #selector(AlbumScreenViewController.searchButton))
         let shareButton = UIBarButtonItem(image: UIImage(systemName: "square.and.arrow.up")?.withRenderingMode(.alwaysTemplate), style: .plain, target: self, action: #selector(AlbumScreenViewController.shareButton))
@@ -56,5 +54,4 @@ class AlbumScreenViewController: UIViewController {
         let flexibleSpace = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: self, action: nil)
         return [searchButton, flexibleSpace, shareButton, flexibleSpace, settingButton, flexibleSpace, informationButton]
     }
-    
 }
