+++ NostelgiAlbum/AlbumScreenViewController/ToolBarController/ShareFunc.swift
