+++ NostelgiAlbum/HomeScreenViewController/HomeDetailController/SetAlbumNameViewController.swift
@@ -2,15 +2,17 @@ import UIKit
 
 // - MARK: SetAlbumNameViewController
 class SetAlbumNameViewController: UIViewController {
-    
+    // MARK: - Properties
     @IBOutlet weak var albumName: UITextField!
     var editVC : HomeEditViewController!
     
+    // MARK: - View Life Cycle
     override func viewDidLoad() {
         albumName.text = editVC.albumName.text
         albumName.becomeFirstResponder()
     }
     
+    // MARK: - Methods
     @IBAction func cancleButton(_ sender: Any) {
         dismiss(animated: true)
     }
@@ -19,7 +21,5 @@ class SetAlbumNameViewController: UIViewController {
         editVC.albumName.text = self.albumName.text
         dismiss(animated: true)
     }
-    
-    
 }
