+++ NostelgiAlbum/HomeScreenViewController/HomeScreenViewController+CollectionView.swift
@@ -1,45 +1,36 @@
 import UIKit
 import RealmSwift
 
-// - MARK: HomeScreenViewController + CollectionView
 extension HomeScreenViewController: UICollectionViewDataSource{
-    
-    // - MARK: CollectionView :: numberOfItemsInSection
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return 3
     }
     
-    // - MARK: CollectionView :: cellForItemAt
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
-        // Cell을 Reuseidenfier를 이용해 생성
+        // Dequeue Reusable cell
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "HomeScreenCollectionViewCell", for: indexPath) as! HomeScreenCollectionViewCell
-        
-        // Cell의 프로퍼티들을 초기화 :: DataReload() 사용 시, cell에 reference로 연결된 객체들을 해제시키기 위함.
+        // Initialize Buttons' properties
         cell.firstButton.setImage(UIImage(systemName: "plus"), for: .normal)
         cell.secondButton.setImage(UIImage(systemName: "plus"), for: .normal)
         cell.firstButton.gestureRecognizers = nil
         cell.secondButton.gestureRecognizers = nil
-        
-        // Button 모양 지정
+        // Buttons' cornerRadius
         cell.firstButton.layer.cornerRadius = 10
         cell.secondButton.layer.cornerRadius = 10
-        
-        // Bottom Label 색 설정
+        // ButtomLabel's backGroundColor
         cell.bottomLabel.backgroundColor = UIColor(red: 0.00, green: 0.00, blue: 0.00, alpha: 1.00)
-        
-        // Cell 배경 색 지정
+        // Cell's backGroundColor
         cell.backgroundColor = UIColor(red: 0.80, green: 0.82, blue: 0.93, alpha: 1.00)
         
-        // CollectionView에 표시되어야 하는 버튼의 개수를 Cover_num 변수에 저장 :: (총 앨범의 개수 + 1)
+        // Number of Cell that should be printed in CollectionView (data.count + 1[Empty Space])
         let cover_num = realm.objects(albumCover.self).count + 1
         
-        // 각 열별로 Case를 나누어 각 열에 존재하는 button들의 isHidden Option을 설정
-        // FIRST CASE :: (cover_num / 2)개 까지의 열을 처리 -> cover_num이 홀수 개인 경우 따로 처리해줘야 함.
+        // MARK: Set cells' Button Properties by indexPath & Number of Data
+        // FIRST CASE :: 해당 row의 cell들이 전부 print 되어야 할 때
         if (indexPath.row + 1) * 2 <= cover_num {
             cell.firstButton.isHidden = false
             cell.secondButton.isHidden = false
-            // First Button :: 무조건 존재 함 -> Set button
+            // First Button
             if let firstbuttonInfo = realm.objects(albumCover.self).filter("id = \(indexPath.row * 2 + 1)").first {
                 if firstbuttonInfo.isCustomCover == false {
                     cell.firstButton.setImage(UIImage(named: firstbuttonInfo.coverImageName), for: .normal)
@@ -54,8 +45,7 @@ extension HomeScreenViewController: UICollectionViewDataSource{
             } else {
                 print("RealmDB Error occur!")
             }
-            // Second Button :: case1. 존재하는 경우 -> Set button
-            // Second Button :: case2. 존재하지 않는 경우 -> Doesn't Set button
+            // Second Button
             if let secondbuttonInfo = realm.objects(albumCover.self).filter("id = \(indexPath.row * 2 + 2)").first {
                 
                 if secondbuttonInfo.isCustomCover == false {
@@ -81,7 +71,7 @@ extension HomeScreenViewController: UICollectionViewDataSource{
             cell.secondButton.isHidden = true
         }
         
-        // 각 Cell의 firstButton / SecondButton Action을 지정
+        // MARK: Set cell's button action properties
         // Callback1 :: First button Action
         cell.callback1 = {
             if cell.firstButton.imageView?.image == UIImage(systemName: "plus"){
@@ -89,8 +79,7 @@ extension HomeScreenViewController: UICollectionViewDataSource{
                 editVC.collectionViewInHome = self.collectionView
                 editVC.modalPresentationStyle = .overCurrentContext
                 self.present(editVC, animated: false)
-            }
-            else {
+            } else {
                 let pushVC = self.storyboard?.instantiateViewController(withIdentifier: "AlbumScreenViewController") as! AlbumScreenViewController
                 pushVC.pageNum = 0
                 pushVC.coverIndex = indexPath.row * 2 + 1
@@ -105,8 +94,7 @@ extension HomeScreenViewController: UICollectionViewDataSource{
                 editVC.collectionViewInHome = self.collectionView
                 editVC.modalPresentationStyle = .overCurrentContext
                 self.present(editVC, animated: false)
-            }
-            else{
+            } else {
                 let pushVC = self.storyboard?.instantiateViewController(withIdentifier: "AlbumScreenViewController") as! AlbumScreenViewController
                 pushVC.pageNum = 0
                 pushVC.coverIndex = indexPath.row * 2 + 2
@@ -114,24 +102,19 @@ extension HomeScreenViewController: UICollectionViewDataSource{
                 self.navigationController?.pushViewController(pushVC, animated: false)
             }
         }
+        // return cell
         return cell
     }
-    
 }
 
-// - MARK: HomeScreenViewController의 UICollectionView Layout을 지정
 extension HomeScreenViewController: UICollectionViewDelegateFlowLayout{
-    
-    // - MARK: 각 Cell별로 width와 Height를 정하는 함수
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         return CGSize(width: collectionView.bounds.width, height: collectionView.bounds.height / 3)
     }
     
-    // - MARK: 각 Cell 사이에 거리를 지정하는 함수
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
         return 0
     }
-    
 }
 
 extension UINavigationController {
