+++ NostelgiAlbum/HomeScreenViewController/SharingAlbumPush.swift
@@ -1,22 +1,15 @@
 import UIKit
 
-// - MARK: shareView Push
 extension HomeScreenViewController {
-    
-    // - MARK: ShareViewController를 만들어 navigationController에 push하는 함수
     func pushShareView(path: URL) {
-        
-        // Share View Controller 생성
+        // Initialize shareViewController
         let shareVC = self.storyboard?.instantiateViewController(withIdentifier: "ShareViewController") as! ShareViewController
-        
-        // nost 파일의 path를 shareVC의 filePath 프로퍼티에 저장
+        // .nost file's path
         shareVC.filePath = path
-        
-        // HomeScreenViewController의 collectionView Object를 shareVC의 프로퍼티에 저장
+        // Set collectionView in shareVC properties to reload Data
         shareVC.collectionViewInHome = collectionView
         
-        // shareVC를 push
-        //self.navigationController?.pushViewController(shareVC, animated: true)
+        // Push shareVC
         shareVC.modalPresentationStyle = .overFullScreen
         self.present(shareVC, animated: true)
     }
