+++ NostelgiAlbum/Supporting Files/Assets.xcassets/AppIcon.appiconset/1024.png
