+++ NostelgiAlbum/Supporting Files/Assets.xcassets/Contents.json
