+++ NostelgiAlbum/Supporting Files/Assets.xcassets/coverImage/Contents.json
