+++ NostelgiAlbum/Supporting Files/Assets.xcassets/coverImage/Pupple.imageset/Contents.json
