+++ NostelgiAlbum/Supporting Files/Assets.xcassets/coverImage/Pupple.imageset/Pupple.png
