+++ NostelgiAlbum/Supporting Files/Assets.xcassets/coverImage/Red.imageset/Contents.json
