+++ NostelgiAlbum/Supporting Files/Assets.xcassets/coverImage/Turquoise.imageset/Contents.json
