+++ NostelgiAlbum/Supporting Files/Assets.xcassets/coverImage/Turquoise.imageset/Turquoise.png
